refact: add useNicknameCheck custom hook
